test: Add small miscellaneous tests
This patch extends various packages and integration tests, increasing
test coverage. They're small enough that it's not worth splitting them
up, as it would add a lot of noise to the history.

+++ test/t-01-simple_local/run.sh
@@ -5,6 +5,10 @@ set -e
 
 init
 
+if ! chasquid --version > /dev/null; then
+	fail "chasquid --version failed"
+fi
+
 # This should fail, as it has no certificates.
 rm -f config/certs/testserver/*.pem
 if chasquid -v=2 --logfile=.logs/chasquid.log --config_dir=config; then
